requirements.txt
flask==2.2.2
gunicorn==20.1.0
uvicorn==0.18.3
pymongo==4.3.3
motor==3.1.1
python-dotenv==0.21.0
requests==2.28.1
aiohttp>=3.9.0
regex==2022.9.13
Werkzeug==2.2.2
pytest==7.2.0
pytest-asyncio==0.20.2
redis==4.3.4
google-generativeai>=0.3.0
nltk==3.8.1
flask[async]